% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_check.cosimmr_output.R
\name{convergence_check}
\alias{convergence_check}
\title{Creates a plot of mean lower bound to check the FFVB model has finished. Uses
 \code{\link{cosimmr_ffvb}} output.}
\usage{
convergence_check(object, ...)
}
\arguments{
\item{object}{An object of class \code{cosimmr_output} produced by the
function \code{\link{cosimmr_ffvb}}}

\item{...}{Not used}
}
\value{
A plot which should show that the mean LB is settling on a value.
}
\description{
Creates diagnostic plot  for an object created with  \code{\link{cosimmr_ffvb}}.
}
\examples{
\donttest{
# A simple example with 10 observations, 2 tracers and 4 sources

# The data
data(geese_data_day1)
cosimmr_1 <- with(
  geese_data_day1,
  cosimmr_load(
    formula = mixtures ~ 1,
    source_names = source_names,
    source_means = source_means,
    source_sds = source_sds,
    correction_means = correction_means,
    correction_sds = correction_sds,
    concentration_means = concentration_means
  )
)

# Plot
plot(cosimmr_1)


# FFVB run
cosimmr_1_out <- cosimmr_ffvb(cosimmr_1)

# Summarise
convergence_check(cosimmr_1_out)
}
}
\seealso{
See \code{\link{cosimmr_ffvb}}for creating objects suitable for 
this function, and many more examples.
See also \code{\link{cosimmr_load}} for creating cosimmr objects,
\code{\link{plot.cosimmr_input}} for creating isospace plots,
\code{\link{plot.cosimmr_output}} for plotting output.
}
\author{
Emma Govan <emmagovan@gmail.com> Andrew Parnell
}
