% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosimmr.R
\name{cosimmr}
\alias{cosimmr}
\alias{<cosimmr>-package}
\title{cosimmr: An R package for Stable Isotope Mixing Models}
\description{
cosimmr is a package that has been developed to allow for running of Stable 
Isotope Mixing Models in R. It allows for the inclusion of covariates and 
has been designed to be easy to use for non-expert users. cosimmr uses Fixed
Form Variational Bayes to run SIMMs, instead of MCMC. This allows for faster
running of models without any issues with convergence
}
\author{
Emma Govan <emmagovan@gmail.com>, Andrew Parnell
}
