% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cosimmr_output.R,
%   R/plot.cosimmr_pred_out.R
\name{plot.cosimmr_output}
\alias{plot.cosimmr_output}
\title{Plot different features of an object created from  \code{\link{cosimmr_ffvb}}.}
\usage{
\method{plot}{cosimmr_output}(
  x,
  type = c("beta_histogram", "beta_boxplot", "p_ind", "p_mean"),
  individuals = 1,
  covariates = 1,
  binwidth = 0.1,
  alpha = 0.5,
  title = NULL,
  n_output = 3600,
  ...
)

\method{plot}{cosimmr_output}(
  x,
  type = c("beta_histogram", "beta_boxplot", "p_ind", "p_mean"),
  individuals = 1,
  covariates = 1,
  binwidth = 0.1,
  alpha = 0.5,
  title = NULL,
  n_output = 3600,
  ...
)
}
\arguments{
\item{x}{An object of class \code{cosimmr_output} created via
\code{\link{cosimmr_ffvb}}.}

\item{type}{The type of plot required. Can be one or more of 'isospace', 
'betahistogram'}

\item{binwidth}{The width of the bins for the histogram. Defaults to 0.05}

\item{alpha}{The degree of transparency of the plots. Not relevant for
matrix plots}

\item{title}{The title of the plot.}

\item{...}{Currently not used}

\item{ggargs}{Extra arguments to be included in the ggplot (e.g. axis limits)}
}
\description{
This function allows for 4 different types of plots of the simmr output
created from \code{\link{cosimmr_ffvb}}. The 
types are: plot of beta values

This function allows for 4 different types of plots of the simmr output
created from \code{\link{cosimmr_ffvb}}. The 
types are: plot of beta values
}
\details{
The matrix plot should form a necessary part of any SIMM analysis since it
allows the user to judge which sources are identifiable by the model.
Further detail about these plots is provided in the vignette.

The matrix plot should form a necessary part of any SIMM analysis since it
allows the user to judge which sources are identifiable by the model.
Further detail about these plots is provided in the vignette.
}
\examples{

\dontrun{
# A simple example with 10 observations, 2 tracers and 4 sources

# The data
data(geese_data_day1)

# Load into simmr
simmr_1 <- with(
  geese_data_day1,
  cosimmr_load(
    formula = mixtures ~ c(1,2,3,3,2,3,2,1,2),
    source_names = source_names,
    source_means = source_means,
    source_sds = source_sds,
    correction_means = correction_means,
    correction_sds = correction_sds,
    concentration_means = concentration_means
  )
)
# Plot
plot(simmr_1)


# FFVB run
simmr_1_out <- cosimmr_ffvb(simmr_1)

plot(simmr_1_out, type = c("isospace", "beta_hist"))
}

\dontrun{
# A simple example with 10 observations, 2 tracers and 4 sources

# The data
data(geese_data_day1)

# Load into simmr
simmr_1 <- with(
  geese_data_day1,
  cosimmr_load(
    formula = mixtures ~ c(1,2,3,3,2,3,2,1,2),
    source_names = source_names,
    source_means = source_means,
    source_sds = source_sds,
    correction_means = correction_means,
    correction_sds = correction_sds,
    concentration_means = concentration_means
  )
)
# Plot
plot(simmr_1)


# FFVB run
simmr_1_out <- cosimmr_ffvb(simmr_1)

plot(simmr_1_out, type = c("isospace", "beta_hist"))
}
}
\seealso{
See  \code{\link{cosimmr_ffvb}} for 
creating objects suitable for this function, and many more examples. See 
also \code{\link{cosimmr_load}} for creating simmr objects, 
\code{\link{plot.cosimmr_input}} for creating isospace plots.

See  \code{\link{cosimmr_ffvb}} for 
creating objects suitable for this function, and many more examples. See 
also \code{\link{cosimmr_load}} for creating simmr objects, 
\code{\link{plot.cosimmr_input}} for creating isospace plots.
}
\author{
Andrew Parnell <andrew.parnell@mu.ie>
}
