% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simmr_data_1}
\alias{simmr_data_1}
\title{Title}
\format{
A list etc with elements
\describe{
\item{mixtures}{what this is and delta 13C etc etc}
\item{source_names}{character vector}
\item{tracer_names}{character vector}
\item{source_means}{character vector}
\item{source_sds}{character vector}
\item{correction_means}{character vector}
\item{correction_sds}{character vector}
\item{concentration_means}{character vector}

}
}
\usage{
simmr_data_1
}
\description{
description
}
\keyword{datasets}
